fix option trade card
