+++ public/css/font.css
@@ -1,35 +1,29 @@
 @font-face {
-  font-family: 'Roboto';
+  font-family: 'SF Pro';
   font-style: normal;
   font-weight: 300;
   font-display: swap;
-  src: url('../fonts/Roboto-Light.ttf');
+  src: local('SF Pro Light'), url('../fonts/SFPRODISPLAYLIGHTITALIC.OTF');
 }
 
 @font-face {
-  font-family: 'Roboto';
+  font-family: 'SF Pro';
   font-style: normal;
   font-weight: 400;
   font-display: swap;
-  src: url('../fonts/Roboto-Regular.ttf');
+  src: local('SF Pro Regular'), url('../fonts/SFPRODISPLAYREGULAR.OTF');
 }
 @font-face {
-  font-family: 'Roboto';
+  font-family: 'SF Pro';
   font-style: normal;
   font-weight: 500;
   font-display: swap;
-  src: url('../fonts/Roboto-Medium.ttf');
+  src: local('SF Pro Medium'), url('../fonts/SFPRODISPLAYMEDIUM.OTF');
 }
 @font-face {
-  font-family: 'Roboto';
+  font-family: 'SF Pro';
   font-style: normal;
   font-weight: bold;
   font-display: swap;
-  src: url('../fonts/Roboto-Bold.ttf');
-}
-
-@font-face {
-  font-family: 'Futura PT';
-  font-display: block;
-  src: url('../fonts/FuturaPT-Demi.ttf');
+  src: local('SF Pro Bold'), url('../fonts/SFPRODISPLAYBOLD.OTF');
 }
