+++ src/pages/App.tsx
@@ -154,7 +154,7 @@ export default function App() {
                   <Route exact strict path="/option_creation/:tab" component={OptionCreation} />
                   <Route exact strict path="/option_trading" component={OptionTrade} />
                   <Route exact strict path="/calculator" component={Calculator} />
-                  <Route exact strict path="/option_trading/:optionId" component={OptionTrade} />
+                  <Route exact strict path="/option_trading/:chainId/:optionId" component={OptionTrade} />
                   <Route exact strict path="/option_exercise" component={OptionExercise} />
                   <Route exact strict path="/liquidity/add/:optionTypeIndex" component={Generate} />
                   {/* <Route exact strict path="/redeem" component={Redeem} /> */}
